+++ zigpy/application.py
@@ -16,7 +16,7 @@ import typing
 from typing import Any, TypeVar
 import warnings
 
-from zigpy.backports import nullcontext
+from zigpy.backports.contextlib import nullcontext
 
 if sys.version_info[:2] < (3, 11):
     from async_timeout import timeout as asyncio_timeout  # pragma: no cover
