+++ zigpy/backports/enum.py
@@ -33,31 +33,3 @@ class StrEnum(str, Enum):
         `StrEnum.auto()` behavior will no longer change.
         """
         raise TypeError("auto() is not supported by this implementation")
-
-
-# Taken straight from CPython 3.12
-class nullcontext:
-    """Context manager that does no additional processing.
-
-    Used as a stand-in for a normal context manager, when a particular
-    block of code is only sometimes used with a normal context manager:
-
-    cm = optional_cm if condition else nullcontext()
-    with cm:
-        # Perform operation, using optional_cm if condition is True
-    """
-
-    def __init__(self, enter_result=None):
-        self.enter_result = enter_result
-
-    def __enter__(self):
-        return self.enter_result
-
-    def __exit__(self, *excinfo):
-        pass
-
-    async def __aenter__(self):
-        return self.enter_result
-
-    async def __aexit__(self, *excinfo):
-        pass
